fix(ScheduledTestsScreen): make swipe-to-refresh work outside lazy list item
This commit fixes an issue that caused swipe-to-refresh to not
work outside the bounds of the lazy list items. This issue was
caused because the size of the lazy list was based on the number
of list items.It has been fixed by using the fillMaxSize()
modifier on the LazyColumn.