fix(TakeTestScreen): call moveToNextWorkBook() when popping backstack from workbook screen
This commit fixes an issue that caused the workbook count to display
in-consistent data. Since TestSessionViewModel#moveToNextWorkBook
was called when navigating to the WorkBookScreen, the workbook
count in the app bar was incremented eventhough the user has
not moved on to the next workbook. In order to fix that, the
method now will be called when navigating back to the
ListenToAudio screen.