fix(LoggedInScreen): use popUpTo when navigating to TakeTestScreen
The popBackStack() method was used to navigate back to the ScheduledTestsScreen
from TakeTestScreen. This caused an issue, wherein, the completed test would
still appear in the list, eventhough the test was marked as completed. In order
to remove the item from the list, a manual refresh was needed. In order to fix
this, the ScheduledTests backstack entry will now be popped off when navigating
to the TakeTestsScreen. As a side effect, the popBackStack() cannot be used
to navigate from TakeTestScreen to ScheduledTestsScreen. Instead, the navigate()
function is used to navigate back to the ScheduledTestsScreen. This will cause
a new instance of ScheduledTestsScreen to be recreated, which, in turn, will
cause the list to referesh.